\name{aCGHex}
\docType{data}
\alias{aCGHex}
\alias{aCGH}

\title{DATA FOR VIGNETTE EXAMPLE}

\description{
  This data object contains information needed for running vignette
  example. 
}

\format{An aCGHplus object}

\details{
  This file contains a data object from the R package aCGHplus. aCGHplus
  is a package designed for array comparative genomic hybridization
  experiments. For information on this package and objects that can be
  created with this package please go to the website:
  http://sphhp.buffalo.edu/biostat/research/software/acghplus/index.php. 
  This is a data set of 10 samples. The mapping.info object and log2
  matricies have been altered to only store information for the region
  4q13; aCGH objects tend to be large, we opted to only include data
  needed for running the vignette example.. 
}

\source{http://sphhp.buffalo.edu/biostat/research/software/acghplus/index.php}

\references{http://sphhp.buffalo.edu/biostat/research/software/acghplus/index.php}

\note{This is included for running vignette example}

\keyword{datasets}





