\name{spheader}

\alias{v1.header}
\alias{v2.header}
\alias{sp.header}

\title{HEADER INFORMATION FOR HTML FILE}

\description{
  Stores header information, tooltip information, for html file
}

\format{Character vector containing lines of html header information}

\details{
  This file contains header information for an html file. It also
  contains javascript tooltip necessary for interactive plot. The
  different versions have different features or work on different web
  browsers.
  v1 works well with firefox and displays information in the upper right
  corner of the web browser
  v2 works well with firefox and displays information at the mouse
  location.
  
}

\source{
  \url{http://www.walterzorn.com/tooltip/tooltip_e.htm}
}

\references{
  \url{http://www.onlamp.com/pub/a/onlamp/2007/07/05/writing-advanced-javascript.html}
  
}

\note{This dataset is used within the sendplot function. There is no
  need for the user to ever call this dataset.}

\keyword{datasets}





