\name{Splot-class}

\docType{class}

\alias{Splot}
\alias{Splot-class}

\title{Splot 'sendplot' Object}

\description{
The Splot object contains all data needed to generate a static layout
of plots, as well as any data mappings for a html file to allow for java
tool-tip display.
}


\author{Lori A. Shepherd, Daniel P. Gaile}

\seealso{\code{\link{initSplot}}, \code{\link{makeImap}}}

\keyword{classes}





