\name{sp.header}

\alias{sp.header}
\alias{spheader}

\title{HEADER INFORMATION FOR HTML FILE}

\description{
  Stores header information for html file
}

\format{Character vector containing 77 lines of html header information}

\details{
  This file contains header information for an html file. It also
  contains javascript necessary for interactive plot.
}

\source{\url{http://www.onlamp.com/pub/a/onlamp/2007/07/05/writing-advanced-javascript.html}}

\references{\url{http://www.onlamp.com/pub/a/onlamp/2007/07/05/writing-advanced-javascript.html}}

\note{This dataset is used within the sendplot function. There is no
  need for the user to ever call this dataset.}

\keyword{datasets}





